misc fix attempts
Tighten up host/guest VMCS state (especially host RIP/RSP and SYSENTER/EFER fields).

Add heavy pre-VMLAUNCH diagnostics (guest+host validation + EPT translation check).

Experiment with a “safe” 32-bit non-paged guest stub (resume_from_virtualization) instead of jumping directly back into UEFI code.

Improve / refactor EPT-based hypervisor hiding, including a new API that can exclude critical code pages.

Clean up some small type/bit handling issues (MSR bitmap, descriptors) and align UEFI feature flags.

+++ hypervisor/src/intel/bitmap.rs
@@ -22,7 +22,7 @@ pub enum MsrOperation {
 
 /// Represents the MSR Bitmap structure used in VMX.
 ///
-/// In processors that support the 1-setting of the “use MSR bitmaps” VM-execution control,
+/// In processors that support the 1-setting of the "use MSR bitmaps" VM-execution control,
 /// the VM-execution control fields include the 64-bit physical address of four contiguous
 /// MSR bitmaps, which are each 1-KByte in size.
 ///
@@ -68,7 +68,7 @@ impl MsrBitmap {
     pub fn modify_msr_interception(&mut self, msr: u32, access: MsrAccessType, operation: MsrOperation) {
         let msr_low = msr & 0x1FFF;
         let msr_index = (msr_low >> 3) as usize;
-        let msr_bit = (msr_low & 7) as u8;
+        let msr_bit = (msr_low & 7) as usize;
 
         let bitmap_section = match (msr >= 0xC000_0000, access) {
             (true, MsrAccessType::Write) => &mut self.write_high_msrs,
@@ -78,8 +78,8 @@ impl MsrBitmap {
         };
 
         match operation {
-            MsrOperation::Hook => bitmap_section[msr_index].set_bit(msr_bit as usize, true),
-            MsrOperation::Unhook => bitmap_section[msr_index].set_bit(msr_bit as usize, false),
+            MsrOperation::Hook => bitmap_section[msr_index].set_bit(msr_bit, true),
+            MsrOperation::Unhook => bitmap_section[msr_index].set_bit(msr_bit, false),
         }
     }
 }
